+++ src/instructor-toolbar/masquerade-widget/index.js
