+++ src/tab-page/LoadedTabPage.jsx
@@ -4,7 +4,7 @@ import PropTypes from 'prop-types';
 import { Header, CourseTabsNavigation } from '../course-header';
 import { useModel } from '../generic/model-store';
 import { useEnrollmentAlert } from '../alerts/enrollment-alert';
-import InstructorToolbar from '../courseware/course/InstructorToolbar';
+import InstructorToolbar from '../instructor-toolbar';
 
 function LoadedTabPage({
   activeTabSlug,
